refactor : 역할 및 책임 분리
- 수식, 연산자, 피연산자 객체 추가
- 문자열 분리 정책, 연산자/피연산자 추출 정책 분리
- 패키지 구분
