--- src/main/kotlin/stringcalculator/domain/OperandExtractor.kt
@@ -1,8 +1,0 @@
-package stringcalculator.domain
-
-class OperandExtractor {
-    fun extract(expString: List<String>): List<Operand> =
-        expString.filterIndexed { index, _ -> index % 2 == 0 }
-            .map { Operand(it.toDouble()) }
-            .toList()
-}
