+++ src/main/kotlin/stringcalculator/domain/operand/Operand.kt
@@ -1,4 +1,4 @@
-package stringcalculator.domain
+package stringcalculator.domain.operand
 
 class Operand(
     val value: Double
