+++ src/test/kotlin/stringcalculator/StringCalculatorTest.kt
@@ -6,7 +6,7 @@ import io.kotest.matchers.should
 import io.kotest.matchers.shouldBe
 import io.kotest.matchers.string.startWith
 import stringcalculator.constants.ErrorType.*
-import stringcalculator.domain.StringCalculator
+import stringcalculator.domain.calculator.StringCalculator
 
 class StringCalculatorTest : BehaviorSpec({
     Given("두 개의 숫자를 더하는 식을 문자열로 제공한다.") {
